refactor(watched): rename MarkAsWatchedDrawer to MarkAsWatchedDrawerHost
The component owns useMarkAsWatched data fetching before rendering the
<Drawer>; align with the host-pattern convention.
