+++ projects/client/src/lib/sections/media-actions/mark-as-watched/MarkAsWatchedDrawerProvider.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-  import MarkAsWatchedDrawer from "./_internal/MarkAsWatchedDrawer.svelte";
+  import MarkAsWatchedDrawerHost from "./_internal/MarkAsWatchedDrawerHost.svelte";
   import { markAsWatchedDrawerStore } from "./_internal/markAsWatchedDrawerStore";
 </script>
 
 {#if $markAsWatchedDrawerStore?.isOpen}
-  <MarkAsWatchedDrawer
+  <MarkAsWatchedDrawerHost
     onClose={markAsWatchedDrawerStore.close}
     title={$markAsWatchedDrawerStore.title}
     {...$markAsWatchedDrawerStore.mediaStore}
